+++ apps/api/src/services/data-provider/data-provider.service.ts
@@ -6,11 +6,7 @@ import {
   IDataProviderResponse
 } from '@ghostfolio/api/services/interfaces/interfaces';
 import { PrismaService } from '@ghostfolio/api/services/prisma.service';
-import {
-  DATE_FORMAT,
-  isGhostfolioScraperApiSymbol,
-  isRakutenRapidApiSymbol
-} from '@ghostfolio/common/helper';
+import { DATE_FORMAT } from '@ghostfolio/common/helper';
 import { Granularity } from '@ghostfolio/common/types';
 import { Injectable } from '@nestjs/common';
 import { DataSource, MarketData } from '@prisma/client';
@@ -20,8 +16,8 @@ import { AlphaVantageService } from './alpha-vantage/alpha-vantage.service';
 import { GhostfolioScraperApiService } from './ghostfolio-scraper-api/ghostfolio-scraper-api.service';
 import { RakutenRapidApiService } from './rakuten-rapid-api/rakuten-rapid-api.service';
 import {
-  convertToYahooFinanceSymbol,
-  YahooFinanceService
+  YahooFinanceService,
+  convertToYahooFinanceSymbol
 } from './yahoo-finance/yahoo-finance.service';
 
 @Injectable()
@@ -37,53 +33,32 @@ export class DataProviderService {
     this.rakutenRapidApiService?.setPrisma(this.prismaService);
   }
 
-  public async get(
-    aSymbols: string[]
-  ): Promise<{ [symbol: string]: IDataProviderResponse }> {
-    if (aSymbols.length === 1) {
-      const symbol = aSymbols[0];
+  public async get(items: IDataGatheringItem[]): Promise<{
+    [symbol: string]: IDataProviderResponse;
+  }> {
+    const response: {
+      [symbol: string]: IDataProviderResponse;
+    } = {};
 
-      if (isGhostfolioScraperApiSymbol(symbol)) {
-        return this.ghostfolioScraperApiService.get(aSymbols);
-      } else if (isRakutenRapidApiSymbol(symbol)) {
-        return this.rakutenRapidApiService.get(aSymbols);
-      }
-    }
-
-    const yahooFinanceSymbols = aSymbols
-      .filter((symbol) => {
-        return (
-          !isGhostfolioScraperApiSymbol(symbol) &&
-          !isRakutenRapidApiSymbol(symbol)
-        );
-      })
-      .map((symbol) => {
-        return convertToYahooFinanceSymbol(symbol);
-      });
-
-    const response = await this.yahooFinanceService.get(yahooFinanceSymbols);
-
-    const ghostfolioScraperApiSymbols = aSymbols.filter((symbol) => {
-      return isGhostfolioScraperApiSymbol(symbol);
-    });
-
-    for (const symbol of ghostfolioScraperApiSymbols) {
-      if (symbol) {
-        const ghostfolioScraperApiResult =
-          await this.ghostfolioScraperApiService.get([symbol]);
-        response[symbol] = ghostfolioScraperApiResult[symbol];
-      }
-    }
-
-    const rakutenRapidApiSymbols = aSymbols.filter((symbol) => {
-      return isRakutenRapidApiSymbol(symbol);
-    });
-
-    for (const symbol of rakutenRapidApiSymbols) {
-      if (symbol) {
-        const rakutenRapidApiResult =
-          await this.ghostfolioScraperApiService.get([symbol]);
-        response[symbol] = rakutenRapidApiResult[symbol];
+    for (const item of items) {
+      if (item.dataSource === DataSource.ALPHA_VANTAGE) {
+        response[item.symbol] = (
+          await this.alphaVantageService.get([item.symbol])
+        )[item.symbol];
+      } else if (item.dataSource === DataSource.GHOSTFOLIO) {
+        response[item.symbol] = (
+          await this.ghostfolioScraperApiService.get([item.symbol])
+        )[item.symbol];
+      } else if (item.dataSource === DataSource.RAKUTEN) {
+        response[item.symbol] = (
+          await this.rakutenRapidApiService.get([item.symbol])
+        )[item.symbol];
+      } else if (item.dataSource === DataSource.YAHOO) {
+        response[item.symbol] = (
+          await this.yahooFinanceService.get([
+            convertToYahooFinanceSymbol(item.symbol)
+          ])
+        )[item.symbol];
       }
     }
 
@@ -91,7 +66,7 @@ export class DataProviderService {
   }
 
   public async getHistorical(
-    aSymbols: string[],
+    aItems: IDataGatheringItem[],
     aGranularity: Granularity = 'month',
     from: Date,
     to: Date
@@ -115,8 +90,17 @@ export class DataProviderService {
           )}'`
         : '';
 
+    const dataSources = aItems.map((item) => {
+      return item.dataSource;
+    });
+    const symbols = aItems.map((item) => {
+      return item.symbol;
+    });
+
     try {
-      const queryRaw = `SELECT * FROM "MarketData" WHERE "symbol" IN ('${aSymbols.join(
+      const queryRaw = `SELECT * FROM "MarketData" WHERE "dataSource" IN ('${dataSources.join(
+        `','`
+      )}') AND "symbol" IN ('${symbols.join(
         `','`
       )}') ${granularityQuery} ${rangeQuery} ORDER BY date;`;
 
@@ -175,13 +159,24 @@ export class DataProviderService {
   }
 
   public async search(aSymbol: string): Promise<{ items: LookupItem[] }> {
-    const { items } = await this.getDataProvider(
-      <DataSource>this.configurationService.get('DATA_SOURCES')[0]
-    ).search(aSymbol);
+    const promises: Promise<{ items: LookupItem[] }>[] = [];
+    let lookupItems: LookupItem[] = [];
 
-    const filteredItems = items.filter((item) => {
+    for (const dataSource of this.configurationService.get('DATA_SOURCES')) {
+      promises.push(
+        this.getDataProvider(DataSource[dataSource]).search(aSymbol)
+      );
+    }
+
+    const searchResults = await Promise.all(promises);
+
+    searchResults.forEach((searchResult) => {
+      lookupItems = lookupItems.concat(searchResult.items);
+    });
+
+    const filteredItems = lookupItems.filter((lookupItem) => {
       // Only allow symbols with supported currency
-      return item.currency ? true : false;
+      return lookupItem.currency ? true : false;
     });
 
     return {