+++ prisma/migrations/20210916182355_added_data_source_to_market_data/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "MarketData" ADD COLUMN     "dataSource" "DataSource" NOT NULL DEFAULT E'YAHOO';